Common parsing and processing of geozones
